Cluster quota limiter
